Add API for calling StableHLO reference interpreter
